+++ docs/conf.py
@@ -17,7 +17,7 @@
 import os
 import sys
 
-sys.path.insert(0, os.path.abspath('../..'))
+sys.path.insert(0, os.path.abspath("../.."))
 
 from skued import __version__
 
@@ -25,7 +25,7 @@ from skued import __version__
 
 # If your documentation needs a minimal Sphinx version, state it here.
 #
-needs_sphinx = '1.0'
+needs_sphinx = "1.0"
 from datetime import datetime
 import alabaster
 
@@ -34,40 +34,42 @@ year = datetime.now().year
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['alabaster',
-              'sphinx.ext.todo',
-              'sphinx.ext.intersphinx',
-              'sphinx.ext.autodoc',
-			  'sphinx.ext.napoleon',
-			  'sphinx.ext.mathjax',
-              'sphinx.ext.autosummary',
-			  'matplotlib.sphinxext.plot_directive']
+extensions = [
+    "alabaster",
+    "sphinx.ext.todo",
+    "sphinx.ext.intersphinx",
+    "sphinx.ext.autodoc",
+    "sphinx.ext.napoleon",
+    "sphinx.ext.mathjax",
+    "sphinx.ext.autosummary",
+    "matplotlib.sphinxext.plot_directive",
+]
 
-intersphinx_mapping = {'numpy': ('http://docs.scipy.org/doc/numpy/', None)}
+intersphinx_mapping = {"numpy": ("http://docs.scipy.org/doc/numpy/", None)}
 
 napoleon_google_docstring = False
 autosummary_generate = True
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 #
 # source_suffix = ['.rst', '.md']
-source_suffix = '.rst'
+source_suffix = ".rst"
 
 # The master toctree document.
-master_doc = 'index'
+master_doc = "index"
 
 # Releases changelog extension
 releases_release_uri = "https://github.com/LaurentRDC/scikit-ued/tree/%s"
 releases_issue_uri = "https://github.com/LaurentRDC/scikit-ued/issues/%s"
 
 # General information about the project.
-project = 'scikit-ued'
-copyright = '%d Laurent P. René de Cotret' % year
-author = 'Laurent P. René de Cotret'
+project = "scikit-ued"
+copyright = "%d Laurent P. René de Cotret" % year
+author = "Laurent P. René de Cotret"
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -89,10 +91,10 @@ language = None
 # directories to ignore when looking for source files.
 # This patterns also effect to html_static_path and html_extra_path
 exclude_patterns = []
-exclude_trees = ['_build']
+exclude_trees = ["_build"]
 
 # The name of the Pygments (syntax highlighting) style to use.
-pygments_style = 'sphinx'
+pygments_style = "sphinx"
 
 # If true, `todo` and `todoList` produce output, else they produce nothing.
 todo_include_todos = True
@@ -103,35 +105,28 @@ todo_include_todos = True
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'sphinx_rtd_theme'
-html_theme_path = ['_themes', ]
-html_sidebars = {
-    '**': [
-        'about.html',
-        'navigation.html',
-        'searchbox.html'
-    ]
-}
+html_theme = "sphinx_rtd_theme"
+html_theme_path = ["_themes"]
+html_sidebars = {"**": ["about.html", "navigation.html", "searchbox.html"]}
 
 # Everything intersphinx's to Python.
-intersphinx_mapping = {
-    'python': ('https://docs.python.org/3.6', None),
-}
+intersphinx_mapping = {"python": ("https://docs.python.org/3.6", None)}
 
 # Autodoc settings
-autodoc_default_flags = ['members']
-autoclass_content = 'both'
-autodoc_member_order = 'groupwise'
+autodoc_default_flags = ["members"]
+autoclass_content = "both"
+autodoc_member_order = "groupwise"
 
 
 def autodoc_skip_member(app, what, name, obj, skip, options):
-    exclusions = {'__weakref__', '__doc__', '__module__', '__dict__'}
+    exclusions = {"__weakref__", "__doc__", "__module__", "__dict__"}
     exclude = name in exclusions
     return skip or exclude
 
 
 def setup(app):
-    app.connect('autodoc-skip-member', autodoc_skip_member)
+    app.connect("autodoc-skip-member", autodoc_skip_member)
+
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
@@ -139,7 +134,7 @@ def setup(app):
 html_static_path = []
 
 # Suppress the warning about a non-local URI for status shields.
-suppress_warnings = ['image.nonlocal_uri']
+suppress_warnings = ["image.nonlocal_uri"]
 
 # Enable releases 'unstable prehistory' mode.
 releases_unstable_prehistory = True