+++ tests/test_array_utils.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 import unittest
 import numpy as np
-from skued import  (
+from skued import (
     repeated_array,
     mirror,
     cart2polar,