+++ tests/time_series/test_nfft.py
@@ -1,4 +1,4 @@
-from skued import  nfft, nfftfreq
+from skued import nfft, nfftfreq
 import unittest
 import numpy as np