+++ tests/time_series/test_time_zero.py
@@ -1,4 +1,4 @@
-from skued import  register_time_shift, register_time_shifts
+from skued import register_time_shift, register_time_shifts
 import unittest
 import numpy as np